# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Gift Shop Catalog Service",
    "description": "Read-only access to an online gift shop's product catalogue, built for a shop\nassistant that helps someone choose a present.\n\nStart with `get_categories` to learn the exact category, occasion and recipient\nvalues this catalogue uses, then use `search_products` for everything else.\n\nEvery operation returns HTTP 200 with a `status` field. `status: \"ok\"` means the\nresult is usable; anything else means nothing was returned and the response\nexplains what to do instead - read `message`, `suggestions` and `notes` and act on\nthem rather than telling the shopper there is nothing. The only error status is\n401, returned when the bearer token is missing or wrong.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://catalog-service-566410667338.europe-west1.run.app",
      "description": "Catalog Service"
    }
  ],
  "paths": {
    "/categories": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "List every category with counts, price ranges, and the exact filter values this catalogue uses.",
        "description": "Returns every category in the gift catalogue - how many products it holds, how many\nare in stock, its price range in euros, and its subcategories. Also returns\n`filter_vocabulary`: the exact strings accepted by the `occasion` and `recipient`\nparameters of the other tools.\n\nCall this once near the start of a gift-finding conversation, before your first\n`search_products` call. It is cheap, and it is the only reliable way to learn the\nexact category, occasion and recipient values this catalogue uses. Guessing them -\n\"homeware\", \"sister\", \"new home\" - returns nothing.\n\nDoes not return individual products. Use `search_products` for those.",
        "operationId": "get_categories",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoriesResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/products/search": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Find gift products matching a shopper's constraints - budget, occasion, recipient, category or free text.",
        "description": "The primary product-discovery tool. Use it as soon as the shopper has given you\n*any* constraint: a budget, an occasion, who it is for, a category, or just a\ndescription of what they want (\"something cosy for a new flat\").\n\nAll parameters are optional and combine with AND. Omit anything the shopper has not\ntold you - do not invent values to fill parameters, because every filter you add\nnarrows the result. Out-of-stock products are excluded unless you explicitly\ninclude them.\n\nReturns at most `limit` products plus `total_matches`, so you know how much you did\nnot see, and `notes` telling you how to narrow a broad result. When nothing\nmatches, `status` is `\"no_match\"` and the response carries concrete suggestions for\nrelaxing the search - act on those rather than telling the shopper there is nothing.\n\nPrefer this over `get_products_by_category` whenever the shopper has said anything\nbeyond naming a category.",
        "operationId": "search_products",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Free text describing what the shopper wants, in their own words where possible - 'cosy blanket for a new flat', 'chef's knife'. Matched against product names, tags, subcategories and descriptions. Leave empty if they have given only structured constraints such as a budget.",
              "nullable": true
            },
            "description": "Free text describing what the shopper wants, in their own words where possible - 'cosy blanket for a new flat', 'chef's knife'. Matched against product names, tags, subcategories and descriptions. Leave empty if they have given only structured constraints such as a budget."
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Category display name ('Home & Living') or slug ('home-living'). Minor misspellings are tolerated. Use a value from get_categories.",
              "nullable": true
            },
            "description": "Category display name ('Home & Living') or slug ('home-living'). Minor misspellings are tolerated. Use a value from get_categories."
          },
          {
            "name": "occasion",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "One of the values in filter_vocabulary.occasions from get_categories. Any other value matches nothing.",
              "nullable": true
            },
            "description": "One of the values in filter_vocabulary.occasions from get_categories. Any other value matches nothing."
          },
          {
            "name": "recipient",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "One of the values in filter_vocabulary.recipients from get_categories. Use 'anyone' only if the shopper genuinely has nobody specific in mind - most products are tagged 'anyone', so it rarely narrows anything usefully.",
              "nullable": true
            },
            "description": "One of the values in filter_vocabulary.recipients from get_categories. Use 'anyone' only if the shopper genuinely has nobody specific in mind - most products are tagged 'anyone', so it rarely narrows anything usefully."
          },
          {
            "name": "max_price_eur",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "minimum": 0,
              "description": "Upper bound in euros, inclusive. Treat a hard limit as hard: 'under 70', 'max 70', 'no more than 70' means pass exactly 70 and never recommend above it. Only vague phrasing - 'around 50', 'about 50', 'or so' - earns headroom, and then pass about 20% more. If the only good option is over a hard limit, say the price plainly and let the shopper decide; do not quietly widen the budget on their behalf.",
              "nullable": true
            },
            "description": "Upper bound in euros, inclusive. Treat a hard limit as hard: 'under 70', 'max 70', 'no more than 70' means pass exactly 70 and never recommend above it. Only vague phrasing - 'around 50', 'about 50', 'or so' - earns headroom, and then pass about 20% more. If the only good option is over a hard limit, say the price plainly and let the shopper decide; do not quietly widen the budget on their behalf."
          },
          {
            "name": "min_price_eur",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "minimum": 0,
              "description": "Lower bound in euros, inclusive. Use only when the shopper wants something that feels substantial; otherwise omit it.",
              "nullable": true
            },
            "description": "Lower bound in euros, inclusive. Use only when the shopper wants something that feels substantial; otherwise omit it."
          },
          {
            "name": "include_out_of_stock",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Leave false. Recommending something unavailable wastes the shopper's time. Set true only when they ask specifically about an item you already know is out of stock.",
              "default": false
            },
            "description": "Leave false. Recommending something unavailable wastes the shopper's time. Set true only when they ask specifically about an item you already know is out of stock."
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "relevance",
                "price_asc",
                "price_desc",
                "rating"
              ],
              "type": "string",
              "description": "Keep 'relevance' unless the shopper asks for the cheapest, the most expensive, or the best reviewed.",
              "default": "relevance"
            },
            "description": "Keep 'relevance' unless the shopper asks for the cheapest, the most expensive, or the best reviewed."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 12,
              "minimum": 1,
              "description": "How many products to return. Ask for 5 even though you will show two or three: the extra give you room to pick the best fit and to hold a backup ready for when the shopper rejects one.",
              "default": 5
            },
            "description": "How many products to return. Ask for 5 even though you will show two or three: the extra give you room to pick the best fit and to hold a backup ready for when the shopper rejects one."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductListResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/products/{product_id}": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Get the full record for one product, including material, colour, brand and exact stock.",
        "description": "Returns everything in the summary plus brand, colour, material, the occasions it\nsuits, who it is for, its tags, and the exact stock count.\n\nCall it when the shopper asks about a specific product you have already\nrecommended, or when you need a detail you do not have - material, colour, brand -\nto answer a question or justify a recommendation. Do not call it for every search\nresult: the fields returned by `search_products` are enough to recommend from.\n\nIf the product is out of stock, the response also carries up to two in-stock\n`alternatives`, so you can offer a replacement in the same reply instead of going\nback to the shopper empty-handed.",
        "operationId": "get_product_details",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Product identifier as returned by the other tools, e.g. 'HL-003'. Case-insensitive. If you have a product name rather than an id, call search_products first."
            },
            "description": "Product identifier as returned by the other tools, e.g. 'HL-003'. Case-insensitive. If you have a product name rather than an id, call search_products first."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDetailResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CategoriesResponse": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ok",
              "no_match",
              "unknown_category",
              "not_found",
              "invalid_parameter"
            ],
            "description": "Always 'ok' for this operation."
          },
          "categories": {
            "items": {
              "$ref": "#/components/schemas/CategoryInfo"
            },
            "type": "array"
          },
          "filter_vocabulary": {
            "$ref": "#/components/schemas/FilterVocabulary"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "description": "Guidance for your next call. Read it before searching.",
            "nullable": true
          }
        },
        "type": "object",
        "required": [
          "status",
          "categories",
          "filter_vocabulary"
        ]
      },
...
      "ProductDetailResponse": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ok",
              "no_match",
              "unknown_category",
              "not_found",
              "invalid_parameter"
            ],
            "description": "'ok' means `product` is present. 'not_found' means the id does not exist."
          },
          "message": {
            "type": "string",
            "description": "Explanation when status is not 'ok'.",
            "nullable": true
          },
          "product": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ProductDetail"
              }
            ],
            "nullable": true
          },
          "alternatives": {
            "type": "array",
            "items": {
              "type": "object"
            },
            "description": "In-stock products similar to this one. Returned automatically when the requested product is out of stock, so you can offer a replacement in the same reply.",
            "nullable": true
          },
          "did_you_mean": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "description": "Product ids with similar names, when the id was not found.",
            "nullable": true
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "nullable": true
          }
        },
        "type": "object",
        "required": [
          "status"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Send the API token as `Authorization: Bearer <token>`.",
        "scheme": "bearer"
      }
    }
  }
}